+++ command/root.go
@@ -10,20 +10,23 @@ import (
 	"os"
 	"os/signal"
 	"strings"
+	"time"
 
-	"github.com/elastic/go-concert/ctxtool/osctx"
-	"github.com/elastic/go-concert/timed"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"go.uber.org/multierr"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 	"golang.org/x/sys/unix"
 
+	"github.com/elastic/go-concert/ctxtool/osctx"
+	"github.com/elastic/go-concert/timed"
+
+	"github.com/elastic/stream/pkg/log"
 	"github.com/elastic/stream/pkg/output"
 
 	// Register outputs.
 	_ "github.com/elastic/stream/pkg/output/gcppubsub"
+	_ "github.com/elastic/stream/pkg/output/httpserver"
 	_ "github.com/elastic/stream/pkg/output/tcp"
 	_ "github.com/elastic/stream/pkg/output/tls"
 	_ "github.com/elastic/stream/pkg/output/udp"
@@ -43,7 +46,7 @@ func Execute() error {
 }
 
 func ExecuteContext(ctx context.Context) error {
-	logger, err := logger()
+	logger, err := log.NewLogger()
 	if err != nil {
 		return nil
 	}
@@ -72,6 +75,13 @@ func ExecuteContext(ctx context.Context) error {
 	rootCmd.PersistentFlags().StringVar(&opts.GCPPubsubOptions.Subscription, "gcppubsub-subscription", "subscription", "GCP Pubsub subscription name")
 	rootCmd.PersistentFlags().BoolVar(&opts.GCPPubsubOptions.Clear, "gcppubsub-clear", true, "GCP Pubsub clear flag")
 
+	// HTTP output flags.
+	rootCmd.PersistentFlags().DurationVar(&opts.HTTPServerOptions.ReadTimeout, "http-server-read-timeout", 5*time.Second, "HTTP Server read timeout")
+	rootCmd.PersistentFlags().DurationVar(&opts.HTTPServerOptions.WriteTimeout, "http-server-write-timeout", 5*time.Second, "HTTP Server write timeout")
+	rootCmd.PersistentFlags().StringVar(&opts.HTTPServerOptions.TLSCertificate, "http-server-tls-cert", "", "Path to the TLS certificate")
+	rootCmd.PersistentFlags().StringVar(&opts.HTTPServerOptions.TLSKey, "http-server-tls-key", "", "Path to the TLS key file")
+	rootCmd.PersistentFlags().StringArrayVar(&opts.HTTPServerOptions.ResponseHeaders, "http-server-response-headers", []string{"content-type", "application/json"}, "List of headers key-values")
+
 	// Sub-commands.
 	rootCmd.AddCommand(newLogRunner(&opts, logger))
 	rootCmd.AddCommand(newPCAPRunner(&opts, logger))
@@ -89,17 +99,6 @@ func ExecuteContext(ctx context.Context) error {
 	rootCmd.PersistentFlags().VisitAll(setFlagFromEnv)
 
 	return rootCmd.ExecuteContext(ctx)
-}
-
-func logger() (*zap.Logger, error) {
-	conf := zap.NewProductionConfig()
-	conf.EncoderConfig.EncodeTime = zapcore.ISO8601TimeEncoder
-	conf.Level = zap.NewAtomicLevelAt(zap.DebugLevel)
-	log, err := conf.Build()
-	if err != nil {
-		return nil, err
-	}
-	return log, nil
 }
 
 func waitForStartSignal(opts *output.Options, parent context.Context, logger *zap.Logger) error {